Zillow-powered-by-Databricks-AI/requirements.txt
streamlit>=1.30.0
databricks-sdk>=0.20.0
databricks-vectorsearch>=0.40
openai>=1.10.0
mlflow>=2.10.0
scikit-learn>=1.4.0
faker>=22.0.0
pydeck>=0.8.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.18.0
